Transition: remap if views in calling activity became invalid
In activity back,  the calling activity might have re-constructed
UI, the accepted views captured in calling activity  was detached
and no longer valid.  Re-query the calling activity for shared
element in such case.  If the app has reconstructed UI of calling
activity,  it is app's responsibility to assign same transition name
to new view where it desires to transition back to.

b/16853679

Change-Id: Ia6729c6082fbaf54c2ce17f695d5b8d66bd9187d

Former-commit-id: b04b36eaadb2f3e56206d8ae157df06bb89f9866