TIF: Move watch history logging to TvProvider
There are only two kinds of watch events that can happen on the system:
 1. The current TV input session is tuned to a new channel.
 2. The session is released for some reason.
The former indicates the end of the previous log entry, if any, followed by the
start of a new entry. The latter indicates the end of the most recent entry for
the given session. Here the system supplies the database the smallest set of
information only that is sufficient to consolidate the log entries while
minimizing database operations in the system service.

Bug: 14791151, Bug: 15921205, Bug: 15988715, Bug: 16831555
Change-Id: Ifcb1a1273c14ee234919d819d082b1004eb92ebc

Former-commit-id: 7eb75dff7a0fb4b19c3e801cd388483d7d471f41