Resolve invalid Typeface style to default
Setting a textAppearance that specified a fontFamily but not a textStyle
results in a styleIndex of -1 being passed to the Typeface.create()
call. This patch resolves such invalid style indices to the default.

Fix for bug 16880318 "Applying TextAppearance To TextView with custom
font family incorrectly adds italics style on LMP"

Change-Id: I05c9cfc4d76161be21b58abdb4037c51a23cae37

Former-commit-id: ff86bcd08827d1050a8cfd99b521c21848b3f364