Revert "Fix a bug: unable to start activity when AM is killing the process."
This reverts commit 7119db2507037010ba8f65711c93017248bffec2 [formerly 5073145f85c8197ec4387d5e2599442a7291ca9c].

I spoke to Dianne and she said that if this is making a problem go away then it is only hiding it and there is a deeper problem we need to fix.

I think if killedByApp is true then we should set app and app.thread to null. I'll look at it further when I have more time.

Change-Id: I44c29e15ca654666440472eab2d2af1c5af80b9a

Former-commit-id: ca9ba58786314f2196d6ade276af5a851d9d9ef8