Fix a regression in partially updading app widgets.
App widgets can be partially updated where only the actions delta
is delivered to the client. There was a regression where instead
of sending the delta, the merged remote views were sent. This was
problematic because certain actions are ignored when merging to
avoid clutter (specifically showNext and showPrevious), hence these
actions never reached the client.

bug:16844814

Change-Id: I88ab58e610cda24154985165e6d6d93a1baa9c79

Former-commit-id: 3de01b2e3474ac2eaccbcf72b6693bb19d254824