Improve adaptive brightness in very dark rooms.
Added config_screenBrightnessDark to configure the minimum value
that will be used for auto-brightness adjustment.  This value is
expected to be less than unadjusted minimum auto-brightness level
to provide some range for the user to make the screen dimmer
in dark rooms.

This configuration value is set to the lowest possible level
by default (1).  Individual devices may need to override this value
in their framework resource overlay depending on their backlight
characteristics.

Change-Id: I9bd3a2355c65f894dff89aeaf7661cdf38f4a6ee

Former-commit-id: 1bfd0f463e279e3ac3f133c5e0ad11ccfd73eed5