When adding an apptoken skip over removed ones.
App tokens are passed from the activity manager to the window
manager along with a list insertion position. That insertion
position presumes that all removed tokens are no longer in window
manager's list. However, when removal of a token is delayed due
to ongoing animation the insertion position was pointing to the
wrong location. This fix skips over tokens that have been marked
for removal when inserting new app tokens.

Fixes bug 15751591.

Change-Id: Ib484c591e2bba9f46ad8e47d60ef05c7bfda0a12

Former-commit-id: 01f79cf91610ec9f85345ea6eeae50ea2f28578f