Do not require device provisioning to do restore-at-install
Provisioning is a milestone used for gating *backup* operation,
but it's important that restores be possible during the setup
process.  In particular, some applications such as home apps may
be deliberately pushed for install after platform restore, but
before the end of the setup process.  We need to be able to do
install-time restores of such apps.

Bug 17288313

Change-Id: Iaff5d9919e6392b2ca5925be4d63a4116cd11f77

Former-commit-id: 674d3e750133efc4b2d575d806596d1f56bda8d9