Allow arbitrarily-added shared elements.
Bug 17112782

This allows the application developer to set any view as
a shared element so that it will avoid the enter and/or exit
transition. This will allow it to be transitioned along with
the shared element transition as well. It will also allow
the called Activity to return different shared elements than
those passed in.

Change-Id: I7b1f4f029af50dc6e2d5f025062b5669486f6427

Former-commit-id: ce725a47f6eee75a6ef86397e9b9fe471aeb1f84