DO NOT MERGE: Use a linear spline if there's a non-monotonic brightness curve
Some devices may contain things like transflective displays where
they actually want the backlight brightness to decrease in the face
of high ambient light scenarios. In the interest of time, just use a
linear interpolation based on the brightness control points for these
cases but in the future we should consider adding a non-monotonic
cubic spline.

Bug: 15611140
Change-Id: I308ad00943d916866ee8d2507a41b7032e6b524e

Former-commit-id: da3b43633fe86d69aab5cab52826e76e01f18d63