Account for scaling effects when cropping
When a scaled window inherits the crop from an attached animating
window the scaling must be accounted for or the crop will obscure
the scaled window.

In the case of the bug that this CL fixes, the SurfaceView containing
a video was scaled down from 1920x1080 by a factor of 0.562 to fit in
1080 width. The crop applied to the window was 1080 but this was passed
to surfaceflinger which ended up cropping the width to 608 due to the
scaling.

Applying the scaling factor to the crop rectangle fixes this bug.

Fixes bug 16334217.

Change-Id: Iafefe43d3696d9fbff01a3666096348468a41e1a

Former-commit-id: 78505d8d096af2981fb39f1a03f9025fd04add77