Automatically supply a content description for ToolbarActionBar nav
If the developer hasn't set a navigation content description on the
Toolbar assigned to be an action bar or a home-action content
description via the ActionBar interface, use the framework default
"navigate up" string.

Also make sure that the default Up description is supplied in the
screen_toolbar window decor layout and that it is parsed properly in
all toolbars, even if we don't have an icon set during construction.

Bug 17298370

Change-Id: Ie2f9e34f92046d4d4ffb9a07e38fa89581891f7b

Former-commit-id: 12b0321c0b400f105cad40568a1f572020ba6336