Unconditionally move stack to front.
Recents was depending on using moveTaskToFront() to also move the
task's stack to the front. But that only works if the task is
actually in the stack because moveTaskToFront() doesn't do the
stack movement if the task is no longer in the stack. This change
calls ActivityStack.moveToFront() in addition to moveTaskToFront().

Fixes bug 17361027.

Also logs callingUid when starting an activity.

Change-Id: I023f956bafae00f45c58757ff5707ce520f02b55

Former-commit-id: 02a4aa261c28df2466234106d61282289276b462