Clear binder identity when sending window change accessibility events.
We get calls for window changes from the window manager and
as a result send accessibility events for these changes. It
is possible that the call for a windows change from the window
manager comes as a result of an unprivileged client calling
into the window manager. However, the event sending code
is performing security checks.

bug:17364244

Change-Id: Ief016f9dafd13ac35418676817848b3ea3ed4d66

Former-commit-id: 9bf08c7bc1844ec8130b47d0ec3783f374b21c46