Change the CAB to use the up-glyph for closing action modes
In an effort to reduce glyph proliferation, change the "done"
checkmark from the CAB to the up-navigation arrow, placing the CAB
firmly in the current navigation hierarchy. This matches behavior for
expandable action views such as SearchView.

Allow the use of different CAB "done" layouts by theme; the material
layout should not include the "done" text and should use the standard
borderless selectable item background.

Bug 17372188

Change-Id: Icfb3e0bbc6a718e22ab27f9d520da5fe4eb833e7

Former-commit-id: b2536cf473a6fc386eea75ab8c97edf3a1ad323b