Fix a regression in SearchView focusability
A bit of code was removed that re-read the View style attributes
during SearchView construction to extract the focusable attribute, but
with a different default value if it was entirely missing. Restore it,
and lump re-reading the focusable attribute in with SearchView's
styleable filter array instead of using the giant View one for a
second time during inflation.

Bug 17292230

Change-Id: Ic5a2dd2d55893b0b290f1f7a3b6de09b37329245

Former-commit-id: ea4ecd6241dbe3a336b39619fd2f281ddfaae1d4