Material EdgeEffect refinements
Add back auto-recede for EdgeEffect, but with a longer delay. Removing
it caused compatibility issues in existing custom views that never
call onRelease. The longer delay will be a signal that something is
amiss without disrupting the UX too badly.

Recede the glow more quickly on release.

Start translucent and fade to more opaque as the glow grows and is
pulled further. Make the opacity stronger in the extreme cases so the
glow is more prominent.

Equalize glow size across the width and height.

Bug 15146438
Bug 15936708

Change-Id: I020d705906e6225b044d0f5303bafa733a679a66

Former-commit-id: 710c456ddb32fe05e13965183e7018015da52eae