Relax requirements on Activity.requestVisibleBehind
The requirement that the top app be resumed in order to request
background visibility was too strict. In particular when the
background app is pausing the top app will be stopped waiting for
pause to complete. This is an appropriate time for the background
app to request visibility but we were rejecting that request.

Also, there is no need for the top app to have an active thread
except to notify it of the changed state.

Fixes bug 17383876.

Change-Id: I52f910baf6c109565694e053445516e1e5fd1c48

Former-commit-id: aea5ceded8015b45e8372b83a02f184974f52d6e