Fix for b/16400888 - API Review: GridLayout attributes
Column weights specify the relative proportions of space distribution 
within a *row*. With that in mind, I flipped the vertical and horizontal 
axes in the docs - but that was a mistake. This CL flips the docs back. 
The code works the right way around so doesn't need to be modified.

Change-Id: I1a36882eee8d1b9bd7b8ef10e03cbf66de0ac261
Former-commit-id: 1af024c9c84393a9514f2207047d24246b6a7045