Make status bar indicators reflect the network actually in use.
Currently, when we are connected to a wifi network without
Internet connectivity (e.g., a captive portal, or an access
point that's not connected to the Internet), we display the wifi
icon and hide the mobile network type icon. This is incorrect,
because all traffic is in fact going over mobile data. This
could cause nasty surprises if mobile data is metered.

Instead, make the icons reflect the true state of affairs:

- A wifi icon means device traffic goes through wifi.
- A wifi icon with ! means wifi is connected but has no
  Internet and traffic is not going through wifi.
- Mobile bars with a network type (e.g., "LTE" means device
  traffic goes through mobile data.
- Mobile bars without a network type mean that mobile data is
  connected but device traffic is not going through mobile data.
- Mobile bars with ! mean that cell is not connected.

Bug: 17237507
Change-Id: I923edbeab3013f10314633c01922c8684591b15c

Former-commit-id: 23dd176915fc118563e510647f048589c12e78e5