Remove auto spaceifying behavior on paste
The paste action contained some logic to automatically insert or remove
spaces, basically to emulate the semantics of the text being a sequence
of space-separated words. This was never universally loved, and gives
particularly unexpected behavior with a hardware keyboard.

This patch simply removes the logic to automatically adjust spaces on
paste, so what gets pasted is simply the literal buffer contents.

Bug: 16900326
Bug: 3304489
Change-Id: I5cdf9c43d98a17aa708883b47245aa7929a124d2

Former-commit-id: 5a689ce78ed85cb5df1f2c7c680dcc6525dfe9eb