Manage mode shouldn't block for other providers.
When loading DocumentsUI in management mode (for Downloads), we
normally focus on a single authority and let the other providers
keep loading async.  However, when one of the providers kicks a
change notification, we end up blocking the main thread.

Simple fix is to move the blocking into the AsyncTask.

Bug: 17452691
Change-Id: Ie5212ad825b7df719690b29941c9ae9ba2cabe4b

Former-commit-id: c9d7150680e288f848178fc6c44776c61ddc8548