isValidSlotId() and isValidPhoneId() updated to handle negatives.
The old version of the code wanted to just check to make sure
that the slot/phone id is greater than INVALID_SLOT_ID or
INVALID_PHONE_ID that are both currently defined to be -1000. The
changes are made for 2 reasons.  First, INVALID_SLOT_ID and/or
INVALID_PHONE_ID might not always be defined to be a negative value
so there is a big assumption.  Secondly, anything greater than
-1000 allows other negative values to be considered valid ids.

Bug: 17243556
Change-Id: I2bbfcc2cfd2d7c4772dfb9c50af2dc88c0f315e2

Former-commit-id: b78890c159f9f06617712261e6c9f284841f2357