Track enable/disable of transport components
For fallback / rollback of backup transport selection we need to
handle live enable/disable of legacy or superseded transports.
We now watch for component enable state changes in packages that
host transports, and rebind as needed.

The semantics for selecting the current transport have also been
adjusted.  We no longer require that the selected transport be
live and currently bound in order to be designated as the active
instance.  This prevents nondeterministic races around upgrade
and replacement.

Bug 17136637

Change-Id: Idaf45cf4522a23576444e6b11626ee3f7f47c36c

Former-commit-id: 64f10efab72aa7701a8d292425f4f56f04ca556b