Customized ripple animation for navigation bar
To implement the new gel-cap shaped ripple I resurrected the
onDraw and setPressed methods from the KitKat implementation.
The KitKat animation timings were adjusted to match the material
ripple, but the structure of the animators is the same as KitKat.

Since the new ripple is drawn directly to the canvas, the references
to the KitKat background image were removed.

The LogInterpolator is copied from the framework Ripple
to match the material animation curves.

Bug: 17112935
Change-Id: If6a3eb92de794b526338166bf8cb096eb9764cf6

Former-commit-id: 42310965fa2c0d2c91bea0a76730a21f6dd308a2