Avoid invalidating view during some outline invalidations
We were invalidating the view during outline invalidation far
too often. Sometimes, it's appropriate to invalidate the view when
the outline changes, but not when the outline is being invalidated
in the middle of drawing the view, which is done in reaction to a view
invalidation.

Issue #17460940 Volantis: panning in Calendar from month to month is only at 30fps

Change-Id: I7bdee17dc9c7df89d4cd3d781010823cef931222

Former-commit-id: 27652eff447a3331701466cf969a50cc106a9940