Missing or wrong pages in preview when scorlling large docs.
The bound pages in the adapter were not properly tracked, more
specifically we did not remove pages from the bound set ending
up in a bad state. Also when the printer changes and if the
new one has different page size the content disappears due to
us not notifying for the data set change.

bug:17515670

Change-Id: Iac74e864609012f6804584fa133e87c27dc186a5

Former-commit-id: f3f963b0bebea91b17f7e60d9b826c458bfde38c