TIF: Re-connect the session on package update.
When TIF client tries to connect a session while TV input is being
updated, updateServiceConnectionLocked() may fail to bindServiceAsUser()
and the session state may remain indefinitely until a client tries to
create another session to connect the service.  Reconnect the session by
calling updateServiceConnectionLocked() when package is updated.

Also, remove the session state when client dies before onSessionCreated().
This was causing the stale session in the above scenario (without
reconnection) to be connected to TIS even when client no longer exists.

Bug: 17518751
Change-Id: I5484df0d80c71649d22438521adf182ab59a6ce4

Former-commit-id: 9c6b5b729bd83b1d1e00428f8a76f272b609c97e