Some accessibility events wrongly filtered out (regression).
We added new APIs to allow accessibility services to query all
windows a user can touch. Sometimes the window state change
event arrives before the window manager sent over the new window
state which leads to a case that the app gets the event and
asks for the window and the window is not there. To address this
if we do not have the window, we hold on to the event and
fire it as soon as the window arrives. This logic is correct
except we were wrongly expecting that the window should have
input focus.

bug:17464645

Change-Id: I1ef50ebddeb4416a6c0776b096bb16aee703700c

Former-commit-id: 4c6a4ce03bb5b20103a656948a1932f2894e7f56