Keep existing API behavior that regressed.
If RemoteViews get a null package in the constructor we are
using the context in which they are applied to obtain resources
during inflation. In such a case if we call the getPackage()
method we were getting null, i.e. the package passed in the
contructor, but now we are returning the package if the app
that created the remote views.

bug:17513823
Change-Id: I0c4a8953c8320469a9160dbaae46ce31465b313b

Former-commit-id: b62424432555d8069f24c2a3550c97a06e5b02f6