Accessibility events may be fired even if no services observe them.
If all accessibility services are disabled we turn off event firing
to avoid a lot of work for nothing. The logic to do that had a flaw
and was not turning event firing off in every case it should. We were
tuning event firing if it is on and no service is enabled but it
is possible that a package on the system image is disabled so the
service is enabled just not bound. Arguably we can remove the
serivice from the enabled list but it is technically on the device.
Note that we correctly handle the case of a package being removed.

bug:17332506

Change-Id: I6d6d9cb9cc271116a3d22ed9fd8d7f533dec9a0b

Former-commit-id: 8aeb0912945ae103223b76d81579073e51579efe