Download PAC when set after a diff PAC URL
Fix bug where rather than downloading new PAC file, the system would
just send out broadcast saying proxy has changed.  Now trigger a
download of the file, and let the broadcast be sent when the download
is complete.

It is safe to pull out this port check as it is also performed at the
time of trying to send the broadcast, and when the port is received.
Removing it takes out a chance of any weird race conditions about
setting the PAC and changing it before the callback completes.

Bug: 17654639
Change-Id: Ic166d8f42bc388998a5aade147579d3ae1355c77

Former-commit-id: 4bf1d218576fa5721bab589cd5945dbe49f0d117