Accessibility in bad state after using SDK tool uiautomatorviewer.
The UiAutormator tool is a part of the SDK. If it is run while
accessibility is enabled it stops all accessibility services
as it is an accessibility service itself to avoid interference
and when done restores back the accessibility state. The issue
was that the accessibility state is not restored leaving the
device in a bad state.

bug:17662770

Change-Id: I3c4f46fa05c76b874eeffdeb867ef433c3fedf2e

Former-commit-id: dd81183bbea34109e1ae2d75667fe4784be6132f