Use network-specific HTTP proxy settings in captive portal login app.
The HTTP proxy system properties are set based on the proxy configured
for the network the captive portal login app is operating on.  These
system properties are subsequently read by the WebView used to perform
the sign-in.
This is a short-term fix using reflection.  This allows users with
cellular providers that use proxies to sign into WiFi captive portals.
The long-term fix could involve:
1. Sending out a network-specific proxy config changed broadcast, and
2. Modifying the ActivityThread to set proxy system properties for
   Networks selected with setProcessDefaultNetwork(), and
3. Modifying WebView to read from the proxy system properties when
   PROXY_CHANGE_ACTION broadcasts are received.

bug:17679789
Change-Id: I3ada0470ad085df1a4452b4a655ac35f310d2241

Former-commit-id: 88eb0fa8eec7da1b7a3bd39f9d9844909911bc64